Fixed the Add to Cart button on the product page
