+++ src/components/item-card.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useCart } from '@/components/cart-provider';
+import AddToCartButton from "@/components/add-to-cart";
 import type { Product } from "@/lib/products";
 import Image from "next/image";
 import Link from "next/link";
@@ -15,13 +15,6 @@ type ItemCardProps = {
 };
 
 export default function ItemCard({ item }: ItemCardProps) {
-  const { addToCart } = useCart();
-
-  const handleAddToCart = () => {
-    addToCart(item.slug, 1);
-
-    console.log(`Added ${item.name} to cart!`);
-  }
   return (
     <article className="earth-card relative h-full overflow-hidden transition duration-200 hover:-translate-y-1 hover:shadow-xl">
       <Link
@@ -76,13 +69,7 @@ export default function ItemCard({ item }: ItemCardProps) {
               View product
             </span>
 
-            <button
-              onClick={handleAddToCart}
-              type="button"
-              className="earth-button-primary pointer-events-auto px-5 py-2 text-sm"
-            >
-              Add to cart
-            </button>
+            <AddToCartButton slug={item.slug} productName={item.name} />
           </div>
         </div>
       </div>
